cleanup context and split useData hook
